move tests from functional to component
