+++ test/functional/cypress/specs/investments/project-edit-task-spec.js
@@ -1,30 +1,17 @@
-import { capitalize } from 'lodash'
-
 import { investments } from '../../../../../src/lib/urls'
 import { id } from '../../../../sandbox/fixtures/whoami.json'
 
 import {
-  assertFieldInput,
   assertBreadcrumbs,
-  assertFieldTextarea,
-  assertFieldRadiosWithLegend,
   assertPayload,
   assertFlashMessage,
   assertExactUrl,
-  assertFieldDate,
 } from '../../support/assertions'
 import { fillWithNewValue } from '../../support/form-fillers'
 import { clickButton } from '../../../../functional/cypress/support/actions'
 import { taskWithInvestmentProjectFaker } from '../../fakers/task'
-import {
-  OPTION_YES,
-  OPTION_NO,
-  DATE_LONG_FORMAT_3,
-} from '../../../../../src/common/constants'
-import {
-  convertDateToFieldDateObject,
-  format,
-} from '../../../../../src/client/utils/date'
+import { DATE_LONG_FORMAT_3 } from '../../../../../src/common/constants'
+import { format } from '../../../../../src/client/utils/date'
 
 describe('Edit investment project task', () => {
   const investmentProjectTask = taskWithInvestmentProjectFaker()
@@ -63,52 +50,6 @@ describe('Edit investment project task', () => {
           detailsUrl,
         [`Edit task for ${investmentProjectTask.investmentProjectTask.investmentProject.investorCompany.name}`]:
           null,
-      })
-    })
-
-    it('should display form fields with values matching the loaded task', () => {
-      cy.dataTest('field-taskTitle').then((element) => {
-        assertFieldInput({
-          element,
-          label: 'Task title',
-          value: investmentProjectTask.title,
-        })
-      })
-
-      cy.get('[data-test="field-taskDescription"]').then((element) => {
-        assertFieldTextarea({
-          element,
-          label: 'Task description (optional)',
-          hint: 'Add details of the task, especially if you intend to assign it to someone else.',
-          value: investmentProjectTask.description,
-        })
-      })
-
-      cy.get('[data-test="field-customDate"]').then((element) => {
-        assertFieldDate({
-          element,
-          label: 'For example 28 11 2025',
-          value: convertDateToFieldDateObject(investmentProjectTask.dueDate),
-        })
-      })
-
-      cy.get('[data-test="field-taskDueDate"]').then((element) => {
-        assertFieldRadiosWithLegend({
-          element,
-          legend: 'Task due date',
-          optionsCount: 7,
-        })
-      })
-
-      cy.get('[data-test="field-taskRemindersEnabled"]').then((element) => {
-        assertFieldRadiosWithLegend({
-          element,
-          legend: 'Do you want to set a reminder for this task?',
-          optionsCount: 3,
-          value: investmentProjectTask.emailRemindersEnabled
-            ? capitalize(OPTION_YES)
-            : capitalize(OPTION_NO),
-        })
       })
     })
